requirements.txt
fastapi
llama-cpp-python